Simplify recursive filtering boundary handling
In addition to being more readable, this avoids a store-to-load
dependency for the current pixel (by propagating loop dependencies in
registers) while not changing the number of instructions in the inner
loop (on x86_64).

About 10% speedup (of whole image) using the scalar implementation on
Apple M1 Pro (both native and Rosetta 2).