Rename hub functions for XYZ specimen position
The specimen coordinates are in um, but they should follow the direction
of the "steps" coordinates, not the um coordinates exposed by
CXYStageBase (which can be flipped using the TransposeMirrorX/Y
properties). Rename to make this distinction clear (without yet fixing
the flipping).

The Z position does not have a flipping issue but treat consistently.